+++ src/bin/todor/clap_app.rs
@@ -17,7 +17,9 @@ pub fn build_cli() -> App<'static, 'static> {
         (@arg IGNORE: -i --("ignore") +takes_value +multiple "Paths to be ignored.")
         (@arg VERBOSE: -v --("verbose") "Provide verbose output.")
         (@arg CHECK: --("check") "Exits nicely only if no TODO tags are found.")
-        (@arg FORMAT: -f --("format") conflicts_with[DELETE_MODE] +takes_value "Outputs TODOs in specified formats.")
+        (@arg FORMAT: -f --("format") conflicts_with[DELETE_MODE] +takes_value
+            possible_values(&["json", "prettyjson", "markdown"])
+            "Outputs TODOs in specified formats.")
         (@arg DELETE_MODE: -d --("delete") conflicts_with[FORMAT] "Interactive delete mode.")
         (@subcommand init =>
             (about: "Creates example config file")