+++ CHANGELOG.md
@@ -17,9 +17,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - completions for bash, zsh, fish, and powershell
 - formula for `brew` package manager
 - global config support
-- formatted output formats using `--format` flag
+- formatted output formats using `-f` flag
 	- JSON
 	- Pretty JSON
+	- Markdown
 
 ### Changed
 - ignore paths are now entirely handled by `todor` bin